third_party/blink/web_tests/scrollingcoordinator/plugin-with-wheel-handler-expected.txt

This test ensures that a plugin which wants to receive wheel events is included in the non-fast scrollable region.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS nonFastScrollableRects.length is 1
PASS rectToString(nonFastScrollableRects[0]) is "[8, 8, 300, 150]"
PASS successfullyParsed is true

TEST COMPLETE
